+++ tests/test_models.py
@@ -10,7 +10,11 @@ from rasterio.crs import CRS
 from rasterio.errors import NotGeoreferencedWarning
 from rasterio.io import MemoryFile
 
-from rio_tiler.errors import InvalidDatatypeWarning, InvalidPointDataError
+from rio_tiler.errors import (
+    InvalidDatatypeWarning,
+    InvalidFormat,
+    InvalidPointDataError,
+)
 from rio_tiler.models import ImageData, PointData
 
 
@@ -237,11 +241,6 @@ def test_point_data():
     for p in PointData(numpy.zeros((3), dtype="uint16")):
         assert p == 0
 
-    pt = PointData(numpy.zeros((3), dtype="uint16"))
-    with pytest.warns(DeprecationWarning):
-        arr = pt.as_masked()
-    assert numpy.ma.is_mask(arr.mask)
-
     pt1 = PointData(numpy.array([1, 2]))
     pt2 = pt1.apply_expression("b1+b2")
     assert pt1.count == 2
@@ -354,40 +353,6 @@ def test_image_apply_colormap():
     assert im.mask[0, 0] == 0
 
 
-def test_image_from_array():
-    """Create ImageData from arrays."""
-    with pytest.warns(DeprecationWarning):
-        arr = numpy.zeros((1, 256, 256), dtype="uint8")
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert im.mask.all()
-
-        arr = numpy.ma.MaskedArray(numpy.zeros((1, 256, 256), dtype="uint8"))
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert im.mask.all()
-
-        arr = numpy.ma.MaskedArray(
-            numpy.zeros((1, 256, 256), dtype="uint8"),
-            mask=numpy.zeros((256, 256), dtype="uint8"),
-        )
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert im.mask.all()
-
-        mask = numpy.zeros((256, 256), dtype="uint8")  # 0 no masked
-        mask[0:10, 0:10] = 1  # masked
-        arr = numpy.ma.MaskedArray(
-            numpy.zeros((1, 256, 256), dtype="uint8"),
-            mask,
-        )
-        im = ImageData.from_array(arr)
-        assert im.data.shape == (1, 256, 256)
-        assert not im.mask.all()
-        assert im.mask[0, 0] == 0
-        assert im.mask[11, 11] == 255
-
-
 def test_image_from_bytes():
     """Create ImageData from bytes."""
     im = ImageData(numpy.zeros((1, 256, 256), dtype="uint8"))
@@ -495,3 +460,9 @@ def test_imagedata_coverage():
 
     coverage = im.get_coverage_array(poly, cover_scale=1000)
     assert numpy.round(numpy.unique(coverage), decimals=3).tolist() == [0, 0.125, 0.25]
+
+
+def test_image_encoding_error():
+    """Test ImageData error when using bad data array shape."""
+    with pytest.raises(InvalidFormat):
+        ImageData(numpy.zeros((5, 256, 256), dtype="uint8")).render(img_format="PNG")